+++ src/utils/generic.ts
@@ -21,3 +21,14 @@ export const resizeCanvas = (canvas: HTMLCanvasElement) => {
     readjustCanvasForBg(canvas);
   }
 };
+
+/**
+ * Wrapper for the native ``vibrate`` method, only available on mobile browsers.
+ * Vibration duration fixed at 300ms.
+ * @author Daniel Desira
+ */
+export const vibrate = () => {
+  if (navigator.vibrate) {
+    navigator.vibrate(300);
+  }
+};